Clear all lint; make zip length invariants explicit
ruff now passes clean on src and tests. Beyond the mechanical fixes, the
substantive change is strict=True on eleven zip() calls where the two
sequences must be the same length -- class names against probability
columns, commands against confidences, epochs against onsets, annotation
onsets against descriptions. Previously a length mismatch would have
silently truncated to the shorter sequence and produced a plausible wrong
answer; now it raises.

Also replaces three pytest.raises(Exception) with the specific
NotFittedError/ValueError, so those tests can no longer pass for the
wrong reason.

+++ src/bwt/deep/modules.py
@@ -30,7 +30,7 @@ import torch.nn as nn
 def _scaled(reference_samples: int, reference_sfreq: float, sfreq: float,
             minimum: int = 1) -> int:
     """Rescale a kernel length quoted at one sampling rate to another."""
-    return max(minimum, int(round(reference_samples * sfreq / reference_sfreq)))
+    return max(minimum, round(reference_samples * sfreq / reference_sfreq))
 
 
 class _MaxNormConv2d(nn.Conv2d):
